+++ services/ExtendedMessage.js
